added role based authorization

+++ convex/files.ts
@@ -21,7 +21,8 @@ async function hasAccessToOrg(ctx: QueryCtx | MutationCtx, orgId: string) {
   }
 
   const hasAccess =
-    user.orgIds.includes(orgId) || user.tokenIdentifier.includes(orgId)
+    user.orgIds.some((items) => items.orgId === orgId) ||
+    user.tokenIdentifier.includes(orgId)
 
   if (!hasAccess) {
     return null
@@ -120,6 +121,14 @@ export const deleteFile = mutation({
 
     if (!access) {
       throw new ConvexError(`You don't have access to perform this action.`)
+    }
+
+    const isAdmin =
+      access.user.orgIds.find((org) => org.orgId === access.file.orgId)
+        ?.role === 'admin'
+
+    if (!isAdmin) {
+      throw new ConvexError('You have to be an admin to perform this action.')
     }
 
     await ctx.db.delete(args.fileId)
